[iOS] Move wallet related code out of TabBrowserData
This change moves all of the wallet code that was living inside TabBrowserData (providers, dapp store, protocol conformances, etc.) into its own tab helper, currently this helper only holds state but it will become a `TabObserver` once the providers & their associated JS are shifted to Chromium
