+++ ios/brave-ios/Sources/Brave/Frontend/Browser/Helpers/WalletTabHelper.swift
@@ -1,0 +1,593 @@
+// Copyright (c) 2026 The Brave Authors. All rights reserved.
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this file,
+// You can obtain one at https://mozilla.org/MPL/2.0/.
+
+import BraveCore
+import BraveWallet
+import Data
+import Foundation
+import Growth
+import Preferences
+import Shared
+import Web
+import os.log
+
+extension TabDataValues {
+  private struct WalletTabHelperKey: TabDataKey {
+    static var defaultValue: WalletTabHelper?
+  }
+  var wallet: WalletTabHelper? {
+    get { self[WalletTabHelperKey.self] }
+    set { self[WalletTabHelperKey.self] = newValue }
+  }
+}
+
+/// A delegate for wallet related events surfaced by a tab's ``WalletTabHelper``
+protocol WalletTabHelperDelegate: AnyObject {
+  /// Shows a wallet notification (e.g. connect / onboarding) for the given origin
+  func showWalletNotification(_ tab: some TabState, origin: URLOrigin)
+  /// Updates the wallet button displayed in the URL bar based on the tab's current wallet state
+  func updateURLBarWalletButton()
+}
+
+/// Encapsulates the wallet (dapp provider) state associated with a tab, acting as the delegate,
+/// events listener and keyring observer for the tab's Ethereum, Solana and keyring services.
+class WalletTabHelper: NSObject {
+  private(set) weak var tab: (any TabState)?
+  weak var delegate: WalletTabHelperDelegate?
+
+  init(tab: some TabState) {
+    self.tab = tab
+    super.init()
+  }
+
+  deinit {
+    // A number of mojo-powered core objects have to be deconstructed on the same
+    // thread they were constructed
+    var mojoObjects: [Any?] = [
+      _walletEthProvider,
+      _walletSolProvider,
+      _walletKeyringService,
+    ]
+
+    DispatchQueue.main.async {
+      // Reference inside to retain it, supress warnings by reading/writing
+      _ = mojoObjects
+      mojoObjects = []
+    }
+  }
+
+  private var _walletEthProvider: BraveWalletEthereumProvider?
+  private var _walletSolProvider: BraveWalletSolanaProvider?
+  private var _walletKeyringService: BraveWalletKeyringService? {
+    didSet {
+      _walletKeyringService?.addObserver(self)
+    }
+  }
+
+  weak var walletEthProvider: BraveWalletEthereumProvider? {
+    get { _walletEthProvider }
+    set { _walletEthProvider = newValue }
+  }
+
+  weak var walletSolProvider: BraveWalletSolanaProvider? {
+    get { _walletSolProvider }
+    set { _walletSolProvider = newValue }
+  }
+
+  weak var walletKeyringService: BraveWalletKeyringService? {
+    get { _walletKeyringService }
+    set { _walletKeyringService = newValue }
+  }
+
+  var tabDappStore: TabDappStore = .init()
+  var isWalletIconVisible: Bool = false {
+    didSet {
+      delegate?.updateURLBarWalletButton()
+    }
+  }
+}
+
+extension WalletTabHelper: BraveWalletProviderDelegate {
+  func showPanel(withOrigin origin: URLOrigin) {
+    guard let tab else {
+      Logger.module.error("Failing to show Wallet panel due to unavailable tab url origin")
+      return
+    }
+    delegate?.showWalletNotification(tab, origin: origin)
+  }
+
+  public func requestPermissions(
+    _ coinType: BraveWallet.CoinType,
+    accounts: [String],
+    origin: URLOrigin,
+    completion: @escaping RequestPermissionsCallback
+  ) {
+    guard let tab else { return }
+    Task { @MainActor in
+      let permissionRequestManager = WalletProviderPermissionRequestsManager.shared
+
+      if permissionRequestManager.hasPendingRequest(for: origin, coinType: coinType) {
+        completion(.requestInProgress, nil)
+        return
+      }
+
+      let isPrivate = tab.isPrivate
+
+      // Check if eth permissions already exist for this origin and if they don't, ensure the user allows
+      // ethereum/solana provider access
+      let walletPermissions =
+        origin.url.map { Domain.walletPermissions(forUrl: $0, coin: coinType) ?? [] } ?? []
+      if walletPermissions.isEmpty {
+        switch coinType {
+        case .eth:
+          if !Preferences.Wallet.allowEthProviderAccess.value {
+            completion(.internal, nil)
+            return
+          }
+        case .sol:
+          if !Preferences.Wallet.allowSolProviderAccess.value {
+            completion(.internal, nil)
+            return
+          }
+        case .ada:
+          if !Preferences.Wallet.allowCardanoProviderAccess.value {
+            completion(.internal, nil)
+            return
+          }
+        case .fil, .btc, .zec, .dot:
+          // not supported
+          fallthrough
+        @unknown default:
+          completion(.internal, nil)
+          return
+        }
+      }
+
+      guard !isPrivate else {
+        completion(.internal, nil)
+        return
+      }
+      switch coinType {
+      case .eth, .sol, .ada:
+        break  // only eth/sol/ada supported for DApps.
+      default:
+        completion(.internal, nil)
+        return
+      }
+      let allowedAccounts = getAllowedAccounts(coinType, accounts: accounts) ?? []
+      if !allowedAccounts.isEmpty {
+        // account(s) requested are allowed
+        completion(.none, allowedAccounts)
+        return
+      }
+
+      // add permission request to the queue
+      let request = permissionRequestManager.beginRequest(
+        for: origin,
+        accounts: accounts,
+        coinType: coinType,
+        providerHandler: completion,
+        completion: { [weak self] response in
+          switch response {
+          case .granted(let accounts):
+            AppReviewManager.shared.processSubCriteria(for: .walletConnectedDapp)
+            completion(.none, accounts)
+          case .rejected:
+            completion(.none, [])
+          }
+          self?.delegate?.updateURLBarWalletButton()
+        }
+      )
+
+      tabDappStore.latestPendingPermissionRequest = request
+      delegate?.showWalletNotification(tab, origin: origin)
+    }
+  }
+
+  /// Returns the selected account if present in `allowedAccounts`, otherwise returns `allowedAccounts`
+  func filterAllowedAccounts(
+    _ allowedAccounts: [String],
+    selectedAccount: String?
+  ) -> [String] {
+    if let selectedAccount = selectedAccount, allowedAccounts.contains(selectedAccount) {
+      return [selectedAccount]
+    }
+    return allowedAccounts
+  }
+
+  /// Fetches all allowed accounts for the current origin.
+  func getAllowedAccounts(_ type: BraveWallet.CoinType, accounts: [String]) -> [String]? {
+    // For Ethereum, locked status is checked in `EthereumProviderImpl::GetAllowedAccounts`
+    // For Solana, locked status is checked in `SolanaProviderImpl::Connect` before
+    // calling `IsAccountAllowed` delegate method.
+    guard let originURL = tab?.visibleURL?.origin.url,
+      let permittedAccountAddresses = Domain.walletPermissions(forUrl: originURL, coin: type)
+    else {
+      return []
+    }
+    // Filter `accounts` requested to return only those permitted
+    return accounts.filter { address in
+      permittedAccountAddresses.contains(where: { permittedAddress in
+        permittedAddress.caseInsensitiveCompare(address) == .orderedSame
+      })
+    }
+  }
+
+  func isAccountAllowed(_ type: BraveWallet.CoinType, account: String) -> Bool {
+    guard let allowedAccounts = getAllowedAccounts(type, accounts: [account]) else {
+      return false
+    }
+    return allowedAccounts.contains(account)
+  }
+
+  func walletInteractionDetected() {
+    // No usage for iOS
+  }
+
+  func showWalletBackup() {
+    // No usage for iOS
+  }
+
+  func unlockWallet() {
+    // No usage for iOS
+  }
+
+  func showWalletOnboarding(withOrigin origin: URLOrigin) {
+    showPanel(withOrigin: origin)
+  }
+
+  func isTabVisible() -> Bool {
+    guard let tab else { return false }
+    return tab.isVisible
+  }
+
+  func isPermissionDenied(_ type: BraveWallet.CoinType) -> Bool {
+    switch type {
+    case .eth:
+      return !Preferences.Wallet.allowEthProviderAccess.value
+    case .sol:
+      return !Preferences.Wallet.allowSolProviderAccess.value
+    case .ada:
+      return !Preferences.Wallet.allowCardanoProviderAccess.value
+    case .fil, .btc, .zec, .dot:
+      return true
+    @unknown default:
+      return true
+    }
+  }
+
+  func showAccountCreation(_ coin: BraveWallet.CoinType, origin: URLOrigin) {
+    guard let tab else { return }
+    let privateMode = tab.isPrivate
+    guard let keyringService = BraveWallet.KeyringServiceFactory.get(privateMode: privateMode)
+    else {
+      return
+    }
+    Task { @MainActor in
+      // check if we receive account creation request without a wallet setup
+      let isWalletCreated = await keyringService.isWalletCreated()
+      if !isWalletCreated {
+        // Wallet is not setup. User must onboard / setup wallet first.
+        delegate?.showWalletNotification(tab, origin: origin)
+        return
+      }
+
+      let accountCreationRequestManager = WalletProviderAccountCreationRequestManager.shared
+
+      // check if same account creation request exists
+      guard !accountCreationRequestManager.hasPendingRequest(for: origin, coinType: coin)
+      else { return }
+
+      // store the account creation request
+      accountCreationRequestManager.beginRequest(for: origin, coinType: coin) { [weak self] in
+        self?.delegate?.updateURLBarWalletButton()
+      }
+      // show wallet notification
+      delegate?.showWalletNotification(tab, origin: origin)
+    }
+  }
+
+  func isSolanaAccountConnected(_ account: String) -> Bool {
+    tabDappStore.solConnectedAddresses.contains(account)
+  }
+
+  func addSolanaConnectedAccount(_ account: String) {
+    Task { @MainActor in
+      tabDappStore.solConnectedAddresses.insert(account)
+    }
+  }
+
+  func removeSolanaConnectedAccount(_ account: String) {
+    Task { @MainActor in
+      tabDappStore.solConnectedAddresses.remove(account)
+    }
+  }
+
+  func clearSolanaConnectedAccounts() {
+    Task { @MainActor in
+      tabDappStore.solConnectedAddresses = .init()
+      await updateSolanaProperties()
+    }
+  }
+}
+
+extension WalletTabHelper: BraveWalletEventsListener {
+  func emitEthereumEvent(_ event: Web3ProviderEvent) {
+    guard let tab, !tab.isPrivate,
+      Preferences.Wallet.defaultEthWallet.value == Preferences.Wallet.WalletType.brave.rawValue,
+      tab.isWebViewCreated
+    else {
+      return
+    }
+    var arguments: [Any] = [event.name]
+    if let eventArgs = event.arguments {
+      arguments.append(eventArgs)
+    }
+    tab.evaluateJavaScript(
+      functionName: "window.ethereum.emit",
+      args: arguments,
+      contentWorld: EthereumProviderScriptHandler.scriptSandbox,
+      completionHandler: nil
+    )
+  }
+
+  func chainChangedEvent(chainId: String) {
+    guard let tab, !tab.isPrivate else { return }
+
+    Task { @MainActor in
+      // chain change might not apply to this origin when assigning
+      // new default network / nil origin: brave-browser #30344
+      guard let provider = walletEthProvider,
+        case let providerChainId = await provider.chainId(),
+        providerChainId == chainId
+      else { return }
+
+      // Temporary fix for #5404
+      // Ethereum properties have been updated correctly, however, dapp is not updated unless there is a reload
+      // We keep the same as Metamask, that, we will reload tab on chain changes.
+      emitEthereumEvent(.ethereumChainChanged(chainId: chainId))
+      await updateEthereumProperties()
+      // don't need to reload if it's a brave wallet webui
+      if tab.visibleURL?.isWalletWebUIURL == true { return }
+      tab.reload()
+    }
+  }
+
+  func accountsChangedEvent(accounts: [String]) {
+    guard let tab, !tab.isPrivate else { return }
+
+    emitEthereumEvent(.ethereumAccountsChanged(accounts: accounts))
+
+    Task {
+      await updateEthereumProperties()
+    }
+  }
+
+  @MainActor
+  func updateEthereumProperties() async {
+    guard let tab, !tab.isPrivate,
+      let keyringService = BraveWallet.KeyringServiceFactory.get(privateMode: false),
+      Preferences.Wallet.defaultEthWallet.value == Preferences.Wallet.WalletType.brave.rawValue
+    else {
+      return
+    }
+
+    // Turn an optional value into a string (or quoted string in case of the value being a string) or
+    // return `undefined`
+    func valueOrUndefined<T>(_ value: T?) -> String {
+      switch value {
+      case .some(let string as String):
+        return "\"\(string)\""
+      case .some(let value):
+        return "\(value)"
+      case .none:
+        return "undefined"
+      }
+    }
+    guard tab.isWebViewCreated, let provider = walletEthProvider else {
+      return
+    }
+
+    let chainId = await provider.chainId()
+    _ = try? await tab.evaluateJavaScript(
+      functionName: "window.ethereum.chainId = \"\(chainId)\"",
+      contentWorld: EthereumProviderScriptHandler.scriptSandbox,
+      asFunction: false
+    )
+
+    let networkVersion = valueOrUndefined(Int(chainId.removingHexPrefix, radix: 16))
+    _ = try? await tab.evaluateJavaScript(
+      functionName: "window.ethereum.networkVersion = \"\(networkVersion)\"",
+      contentWorld: EthereumProviderScriptHandler.scriptSandbox,
+      asFunction: false
+    )
+
+    let isKeyringLocked = await keyringService.isLocked()
+    let selectedAccount: String
+    if isKeyringLocked {
+      // Check for locked status before assigning account address.
+      // `getAllowedAccounts` is not async, can't check locked status.
+      selectedAccount = valueOrUndefined(Optional<String>.none)
+    } else {
+      let allAccounts = await keyringService.allAccounts()
+      let allEthAccounts = allAccounts.accounts.filter { $0.coin == .eth }
+      if let allowedAccounts = getAllowedAccounts(.eth, accounts: allEthAccounts.map(\.address)) {
+        let selectedAccountForCoin = allAccounts.ethDappSelectedAccount
+        let filteredAllowedAccounts = filterAllowedAccounts(
+          allowedAccounts,
+          selectedAccount: selectedAccountForCoin?.address
+        )
+        selectedAccount = valueOrUndefined(filteredAllowedAccounts.first)
+      } else {
+        selectedAccount = valueOrUndefined(Optional<String>.none)
+      }
+    }
+    _ = try? await tab.evaluateJavaScript(
+      functionName: "window.ethereum.selectedAddress = \(selectedAccount)",
+      contentWorld: EthereumProviderScriptHandler.scriptSandbox,
+      asFunction: false
+    )
+  }
+
+  func messageEvent(subscriptionId: String, result: BaseValue) {
+    let eventArgs = BaseValue(dictionaryValue: [
+      "type": BaseValue(stringValue: "eth_subscription"),
+      "data": BaseValue(dictionaryValue: [
+        "subscription": BaseValue(stringValue: subscriptionId),
+        "result": result,
+      ]),
+    ])
+    emitEthereumEvent(.init("message", arguments: eventArgs.jsonObject))
+  }
+}
+
+extension WalletTabHelper: BraveWalletSolanaEventsListener {
+  func accountChangedEvent(account: String?) {
+    guard let tab else { return }
+    Task { @MainActor in
+      if tab.isWebViewCreated {
+        let script: String
+        if let account = account {
+          script =
+            "if (\(UserScriptManager.walletSolanaNameSpace).solanaWeb3) { window.solana.emit('accountChanged', new \(UserScriptManager.walletSolanaNameSpace).solanaWeb3.PublicKey('\(account.htmlEntityEncodedString)')) }"
+        } else {
+          script = "window.solana.emit('accountChanged')"
+        }
+        _ = try? await tab.evaluateJavaScript(
+          functionName: script,
+          contentWorld: .page,
+          asFunction: false
+        )
+      }
+      await updateSolanaProperties()
+    }
+  }
+
+  func disconnectEvent() {
+    emitSolanaEvent(.disconnect)
+  }
+
+  func emitSolanaEvent(_ event: Web3ProviderEvent) {
+    guard let tab, tab.isWebViewCreated,
+      Preferences.Wallet.defaultSolWallet.value == Preferences.Wallet.WalletType.brave.rawValue
+    else {
+      return
+    }
+    Task {
+      var arguments: [Any] = [event.name]
+      if let eventArgs = event.arguments {
+        arguments.append(eventArgs)
+      }
+      try await tab.evaluateJavaScript(
+        functionName: "window.solana.emit",
+        args: arguments,
+        contentWorld: .page
+      )
+    }
+  }
+
+  @MainActor func updateSolanaProperties() async {
+    guard let tab,
+      Preferences.Wallet.defaultSolWallet.value == Preferences.Wallet.WalletType.brave.rawValue,
+      tab.isWebViewCreated,
+      let provider = walletSolProvider
+    else {
+      return
+    }
+    let isConnected = await provider.isConnected()
+    _ = try? await tab.evaluateJavaScript(
+      functionName: "window.solana.isConnected = \(isConnected)",
+      contentWorld: .page,
+      asFunction: false
+    )
+    // publicKey
+    if let keyringService = walletKeyringService,
+      let publicKey = await keyringService.allAccounts().solDappSelectedAccount?.address,
+      self.isSolanaAccountConnected(publicKey)
+    {
+      _ = try? await tab.evaluateJavaScript(
+        functionName: """
+          if (\(UserScriptManager.walletSolanaNameSpace).solanaWeb3) {
+            window.__firefox__.execute(function($) {
+              window.solana.publicKey = $.deepFreeze(
+                new \(UserScriptManager.walletSolanaNameSpace).solanaWeb3.PublicKey('\(publicKey.htmlEntityEncodedString)')
+              );
+            });
+          }
+          """,
+        contentWorld: .page,
+        asFunction: false
+      )
+    }
+  }
+}
+
+extension WalletTabHelper: BraveWalletKeyringServiceObserver {
+  func walletCreated() {
+  }
+
+  func walletRestored() {
+  }
+
+  func walletReset() {
+    guard let tab else { return }
+    tab.reload()
+    delegate?.updateURLBarWalletButton()
+  }
+
+  func locked() {
+    Task { @MainActor in
+      await updateEthereumProperties()
+      await updateSolanaProperties()
+    }
+  }
+
+  func unlocked() {
+    guard let origin = tab?.visibleURL?.origin,
+      let keyringService = walletKeyringService
+    else { return }
+    Task { @MainActor in
+      // check domain already has some permitted accounts for this Tab's URLOrigin
+      let permissionRequestManager = WalletProviderPermissionRequestsManager.shared
+      let allAccounts = await keyringService.allAccounts().accounts
+      for coin in WalletConstants.supportedCoinTypes(.dapps) {
+        let allAccountsForCoin = allAccounts.filter { $0.coin == coin }
+        if permissionRequestManager.hasPendingRequest(for: origin, coinType: coin) {
+          let pendingRequests = permissionRequestManager.pendingRequests(
+            for: origin,
+            coinType: coin
+          )
+          let accounts = getAllowedAccounts(coin, accounts: allAccountsForCoin.map(\.address)) ?? []
+          if !accounts.isEmpty {
+            for request in pendingRequests {
+              // cancel the requests if `allowedAccounts` is not empty for this domain
+              permissionRequestManager.cancelRequest(request)
+              // let wallet provider know we have allowed accounts for this domain
+              request.providerHandler?(.none, accounts)
+            }
+          }
+        }
+      }
+    }
+  }
+
+  func backedUp() {
+  }
+
+  func accountsChanged() {
+  }
+
+  func autoLockMinutesChanged() {
+  }
+
+  func selectedWalletAccountChanged(account: BraveWallet.AccountInfo) {
+  }
+
+  func selectedDappAccountChanged(coin: BraveWallet.CoinType, account: BraveWallet.AccountInfo?) {
+  }
+
+  func accountsAdded(addedAccounts: [BraveWallet.AccountInfo]) {
+  }
+}
